# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Bank System",
        "version": "0.1.0",
        "description": "This is Bank System API documentation",
        "contact": {
            "name": "Naufal Azmi Ginting",
            "url": "https://github.com/boboiazumi",
            "email": "naufal2404ginting@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        },
        {
            "url": "https://binarkm7.et.r.appspot.com/"
        },
        {
            "url": "http://35.219.74.122/"
        }
    ],
    "paths": {
        "/api/v1/users": {
            "get": {
                "summary": "List all users",
                "description": "List all users",
                "tags": ["users"],
                "responses": {
                    "200": {
                        "description": "The List Of the books",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Anton Wibowo"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "anton@gmail.com"
                                                    },
                                                    "password": {
                                                        "type": "string",
                                                        "example": "$2b$04$9o0rDYDbVrrzpnXBupjf6e2D/KIDWC.7m7qrbMTzyEQ3jV8firisW"
                                                    }
                                                }
                                            },
                                            "example": [
                                                {
                                                    "id": 39,
                                                    "name": "Zaenal Abidin",
                                                    "email": "nadiakomalasari@gmail.com",
                                                    "password": "$2b$04$9o0rDYDbVrrzpnXBupjf6e2D/KIDWC.7m7qrbMTzyEQ3jV8firisW"
                                                },
                                                {
                                                    "id": 40,
                                                    "name": "Achzanu Wibowo",
                                                    "email": "achzanu@gmail.com",
                                                    "password": "$2b$04$JQp//mO4uTladWplMmQ5KuGYf88Iz7qNWJXOVnJymAsh40EujpYKu"
                                                }
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create Users",
                "description": "Create Users",
                "tags": ["users"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "data": {
                                        "type": "object",
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Zaenal Abidin"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "nadiakomalasari@gmail.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "ZaenalAbidin12345"
                                            },
                                            "profile": {
                                                "type": "object",
                                                "properties": {
                                                    "identityType": {
                                                        "type": "string",
                                                        "example": "KTP"
                                                    },
                                                    "identityNumber": {
                                                        "type": "string",
                                                        "example": "1023342764323"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "Alamat Rumah Zaenal Abidin"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success create new user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "SUCCESS"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
		"/api/v1/accounts": {
			"get": {
				"description": "Get all bank account",
                "summary": "Get all bank account",
				"tags": ["accounts"],
				"responses": {
					"200":{
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"status": {"type": "string", "example": "OK"},
										"data": {"type": "array", "items": {
											"properties": {
												"id": {"type": "number", "example": 49},
												"userId": {"type": "number", "example": 49},
												"bankName": {"type": "string", "example": "BANK INDONESIA"},
            									"bankAccountNumber": {"type": "number", "example": 49},
           										"balance": {"type": "number", "example": 5000}
											}
										}}
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"description": "Add new bank account",
                "summary": "Add new bank account",
				"tags": ["accounts"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"data" : {
										"type" :"object",
										"properties": {
											"bankName": {"type":"string", "example": "BANK INDONESIA"}
										}
									}
								}
							}
						}
					}
				},
				"responses": {
					"200":{
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"status": {"type": "string", "example": "OK"}
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/api/v1/transactions": {
			"get": {
				"description": "List of transactions",
				"summary": "List of transactions",
				"tags": ["transactions"],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"status": {"type": "string", "example":"OK"},
										"data": {
											"type": "array",
											"items": {
												"properties": {
													"id": {"type": "number", "example": 6},
													"sourceAccountId": {"type": "number", "example": 49},
													"destinationAccountId": {"type": "number", "example": 50},
													"amount": {"type": "number", "example": 5000}
												}
											}
										}
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"description": "Create new transactions",
				"summary": "Create new transactions",
				"tags": ["transactions"],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"data": {"type": "object", "properties": {
										"sourceBankAccountNumber": {"type": "number", "example": 49},
            							"destinationBankAccountNumber": {"type": "number", "example": 50},
            							"amount": {"type": "number", "example": 1000}
									}}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"status": {"type": "string", "example":"OK"}
									}
								}
							}
						}
					}
				}
			}
		},
...
        "/api/v1/register": {
            "post": {
                "summary": "Register Users",
                "description": "Register Users",
                "tags": ["register"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "data": {
                                        "type": "object",
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Zaenal Abidin"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "nadiakomalasari@gmail.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "ZaenalAbidin12345"
                                            },
                                            "profile": {
                                                "type": "object",
                                                "properties": {
                                                    "identityType": {
                                                        "type": "string",
                                                        "example": "KTP"
                                                    },
                                                    "identityNumber": {
                                                        "type": "string",
                                                        "example": "1023342764323"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "Alamat Rumah Zaenal Abidin"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success create new user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "SUCCESS"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/authenticate": {
            "post": {
                "summary": "Authenticate",
                "description": "Authenticate",
                "tags": ["authenticate"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "data": {
                                        "type": "object",
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "nadiakomalasari@gmail.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "ZaenalAbidin12345"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success create new user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "SUCCESS"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MzksImlhdCI6MTczMDIxNzE4MiwiZXhwIjoxNzMwMzAzNTgyfQ.UKGeoI5TeFt2KXshKssSmLiNY6fUn8XZ-8AL8HXIsBY"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/uploads": {
            "summary": "Need authenticate first",
            "get": {
                "summary": "Get all uploaded image",
                "description": "Get all uploaded image",
                "tags": ["uploads"],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {"type": "string", "example": "OK"},
                                        "data": {"type": "array", "items": {"type": "object", "properties": {
                                            "id": {"type": "integer", "example": 1},
                                            "userId": {"type": "integer", "example": 44},
                                            "title": {"type": "string", "example": "Nahida"},
                                            "description": {"type": "string", "example":"Nahida Pictures"},
                                            "url": {"type":"string","example":"https://ik.imagekit.io/azumidesu/9187d712951e99f926d2a7e018cf171d_a0Y7Z9puP"},
                                            "fileId": {"type":"string", "example":"672d9259e375273f6021ff48"}
                                        }}}
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Upload an image",
                "description": "Upload an image",
                "tags": ["uploads"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {"type": "string", "example": "Pemandangan Indah"},
                                    "description": {"type": "string", "example": "Pemandangan di isekai tahun 2030"},
                                    "img": {"type": "string", "format": "binary"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {"type": "string", "example": "OK"}
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/forgot-password": {
            "post": {
                "summary": "Forgot Password",
                "tags": ["forgot-password"],
                "description": "Send a link to reset the password to email. Note: email must be linked to the user and be a valid email.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {"type": "string", "example":"thed0704@gmail.com"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json":{
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {"type": "string", "example": "SUCCESS"}
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
		"securitySchemes": {
			"bearerAuth":{
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"security": [
		{"bearerAuth": []}
	],
...
}